chore: tanstack setup
